[BE] make StripeSubscriptionPlan.amount a BigInteger

+++ aa_stripe/models.py
@@ -332,7 +332,7 @@ class StripeSubscriptionPlan(StripeBasicModel):
 
     is_created_at_stripe = models.BooleanField(default=False)
     source = JSONField(blank=True, help_text=_("Source of the plan, ie: {\"prescription\": 1}"))
-    amount = models.IntegerField(help_text=_("In cents. More: https://stripe.com/docs/api#create_plan-amount"))
+    amount = models.BigIntegerField(help_text=_("In cents. More: https://stripe.com/docs/api#create_plan-amount"))
     currency = models.CharField(
         max_length=3, help_text=_("3 letter ISO code, default USD, https://stripe.com/docs/api#create_plan-currency"),
         default="USD")
